Add a second definition for the word 'roar'
Instead of using the WORD in the URL:

    /definitions/roar/

We're now going to have to use the raw strokes:

    /definitions/ror/
    /definitions/roer/

This will make it possible for the system to contain more than one
definition for each word.

But we need to be able to translate definitions that are specified as
'bricks' into the raw strokes so that we can generate these URLs.