Add definitions for "wealth" and "death"
Paths:
    /definitions/we*lt
    /definitions/tke*t

I had to check this, but it appears that the '*' character is safe to
use in URLs and does not need to be encoded:

    https://perishablepress.com/stop-using-unsafe-characters-in-urls/

Current version of the SVG generating code draws diagrams in which the
'*T' brick obscures the 'E' and 'L' bricks in these words.

Next step: make SVG generating code detect such collisions and
accommodate them.