Annotate 'lift' and 'lurch' definitions with collisions
The stroke that produces the word 'lurch' could just as well serve as a
stroke for the word 'lunch'. On the page for the 'lurch' stroke, we'll
include a link to the page for the word 'lunch'.

Likewise, the stroke for 'lift' could just as well produce the word
'list', so we'll illustrate this collision too.